requirements.txt
alembic==1.7.7
certifi==2023.5.7
charset-normalizer==3.1.0
flake8==6.0.0
greenlet==2.0.2
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
python-dotenv==1.0.0
requests==2.30.0
schedule==1.2.0
sentry-sdk==1.23.0
SQLAlchemy==1.4.36
typing_extensions==4.5.0
urllib3==1.26.15
